Voting: reorganize hooks into separate files in hooks/
