+++ apps/voting/app/src/components/VoteActions.js
@@ -12,8 +12,8 @@ import {
   useTheme,
 } from '@aragon/ui'
 import { useAppState, useConnectedAccount } from '@aragon/api-react'
+import useExtendedVoteData from '../hooks/useExtendedVoteData'
 import { noop, formatDate } from '../utils'
-import { useExtendedVoteData } from '../vote-hooks'
 import { VOTE_NAY, VOTE_YEA } from '../vote-types'
 import { isVoteAction } from '../vote-utils'
