+++ apps/voting/app/src/components/VoteCasted.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { GU, IconCheck, RADIUS, textStyle, useTheme } from '@aragon/ui'
 import { useAppState } from '@aragon/api-react'
-import { useExtendedVoteData } from '../vote-hooks'
+import useExtendedVoteData from '../hooks/useExtendedVoteData'
 import { VOTE_YEA } from '../vote-types'
 
 function VoteCasted({ vote }) {
