+++ apps/voting/app/src/hooks/useNow.js
@@ -1,0 +1,14 @@
+import { useEffect, useState } from 'react'
+
+export default function useNow(updateEvery = 1000) {
+  const [now, setNow] = useState(new Date())
+  useEffect(() => {
+    const timer = setInterval(() => {
+      setNow(new Date())
+    }, updateEvery)
+    return () => {
+      clearInterval(timer)
+    }
+  }, [updateEvery])
+  return now
+}
