+++ apps/voting/app/src/hooks/usePanelState.js
@@ -1,41 +1,9 @@
-import { useCallback, useEffect, useMemo, useState } from 'react'
-import { noop } from './utils'
-
-export function useNow(updateEvery = 1000) {
-  const [now, setNow] = useState(new Date())
-  useEffect(() => {
-    const timer = setInterval(() => {
-      setNow(new Date())
-    }, updateEvery)
-    return () => {
-      clearInterval(timer)
-    }
-  }, [updateEvery])
-  return now
-}
-
-export function usePromise(fn, memoParams, defaultValue) {
-  const [result, setResult] = useState(defaultValue)
-  useEffect(() => {
-    let cancelled = false
-    const promise = typeof fn === 'function' ? fn() : fn
-    promise.then(value => {
-      if (!cancelled) {
-        setResult(value)
-      }
-    })
-    return () => {
-      cancelled = true
-    }
-
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [...memoParams, fn])
-  return result
-}
+import { useCallback, useMemo, useState } from 'react'
+import { noop } from '../utils'
 
 // Handles the state of a panel.
 // Pass `onTransitionEnd` to the same SidePanel prop.
-export function usePanelState({ onDidOpen = noop, onDidClose = noop } = {}) {
+export default function usePanelState({ onDidOpen = noop, onDidClose = noop } = {}) {
   const [visible, setVisible] = useState(false)
 
   // `didOpen` is set to `true` when the opening transition of the panel has
@@ -70,10 +38,4 @@ export function usePanelState({ onDidOpen = noop, onDidClose = noop } = {}) {
     () => ({ requestOpen, requestClose, visible, didOpen, onTransitionEnd }),
     [requestOpen, requestClose, visible, didOpen, onTransitionEnd]
   )
-}
-
-export function useScrollTop(dependency) {
-  useEffect(() => {
-    window.scrollTo(0, 0)
-  }, [dependency])
 }
