+++ apps/voting/app/src/hooks/useScrollTop.js
@@ -1,0 +1,7 @@
+import { useEffect } from 'react'
+
+export default function useScrollTop(dependency) {
+  useEffect(() => {
+    window.scrollTo(0, 0)
+  }, [dependency])
+}
