--- apps/voting/app/src/vote-hooks.js
@@ -1,105 +1,0 @@
-import { useEffect, useMemo, useState } from 'react'
-import { useAragonApi, useAppState } from '@aragon/api-react'
-import {
-  getCanExecute,
-  getCanVote,
-  getUserBalance,
-  isVoteOpen,
-} from './vote-utils'
-import { useNow, usePromise } from './utils-hooks'
-import { VOTE_ABSENT } from './vote-types'
-import TOKEN_ABI from './abi/token-balanceOfAt.json'
-
-// Get the votes array ready to be used in the app.
-export function useVotes() {
-  const { votes, connectedAccountVotes } = useAppState()
-  const now = useNow()
-
-  const openedStates = (votes || []).map(v => isVoteOpen(v, now))
-  const openedStatesKey = openedStates.join('')
-
-  return useMemo(() => {
-    if (!votes) {
-      return []
-    }
-    return votes.map((vote, i) => ({
-      ...vote,
-      data: {
-        ...vote.data,
-
-        metadata: vote.data.metadata || '',
-        description: vote.data.description || '',
-        open: openedStates[i],
-      },
-      connectedAccountVote: connectedAccountVotes[vote.voteId] || VOTE_ABSENT,
-    }))
-  }, [votes, connectedAccountVotes, openedStatesKey]) // eslint-disable-line react-hooks/exhaustive-deps
-}
-
-// Load and returns the token contract, or null if not loaded yet.
-export function useTokenContract() {
-  const { api, appState } = useAragonApi()
-  const { tokenAddress } = appState
-  const [contract, setContract] = useState(
-    api && tokenAddress ? api.external(tokenAddress, TOKEN_ABI) : null
-  )
-
-  useEffect(() => {
-    // We assume there is never any reason to set the contract back to null.
-    if (api && tokenAddress && !contract) {
-      setContract(api.external(tokenAddress, TOKEN_ABI))
-    }
-  }, [api, tokenAddress, contract])
-
-  return contract
-}
-
-// Get the extended data related to a vote
-export function useExtendedVoteData(vote) {
-  const {
-    api,
-    connectedAccount,
-    appState: { tokenDecimals },
-  } = useAragonApi()
-
-  const tokenContract = useTokenContract()
-
-  const canExecutePromise = useMemo(() => getCanExecute(vote, api), [
-    vote && vote.voteId,
-    api,
-  ])
-  const canExecute = usePromise(
-    canExecutePromise,
-    [vote && vote.voteId, api],
-    false
-  )
-
-  const canUserVotePromise = useMemo(
-    () => getCanVote(vote, connectedAccount, api),
-    [vote && vote.voteId, connectedAccount, api]
-  )
-  const canUserVote = usePromise(
-    canUserVotePromise,
-    [vote && vote.voteId, connectedAccount, api],
-    false
-  )
-
-  const userBalancePromise = useMemo(
-    () => getUserBalance(vote, connectedAccount, tokenContract, tokenDecimals),
-    [vote && vote.voteId, connectedAccount, tokenContract, tokenDecimals]
-  )
-  const userBalance = usePromise(
-    userBalancePromise,
-    [vote && vote.voteId, connectedAccount, tokenContract, tokenDecimals],
-    -1
-  )
-
-  return {
-    canExecute,
-    canUserVote,
-    userBalance,
-    userBalancePromise,
-    canUserVotePromise,
-    canExecutePromise,
-  }
-}
